+++ pixelcity-wayland/Random.h
@@ -1,5 +1,6 @@
+#include <stdint.h>
 #define COIN_FLIP     (RandomVal (2) == 0)
 
-unsigned long RandomVal (int range);
-unsigned long RandomVal (void);
-void          RandomInit (unsigned long seed);
+uint32_t RandomVal (int range);
+uint32_t RandomVal (void);
+void          RandomInit (uint32_t seed);
